finally I am able to perform some competition
